[rollout] refactor: move seed helpers to agent_loop.utils
Extract rollout seed helpers into utils.py with short docstrings, drop
data.seed rollout fallback, and fix OCR recipe comments outside the
continued python command so the script stays runnable.
